Change the protocol names
