--- Yoshi/QAKit/EnvironmentMenu/Environment.swift
@@ -1,23 +1,0 @@
-//
-//  YoshiEnvironment.swift
-//  Yoshi
-//
-//  Created by Kanglei Fang on 28/08/2017.
-//  Copyright © 2017 Prolific Interactive. All rights reserved.
-//
-
-import Foundation
-
-internal func == (left: Environment , right: Environment) -> Bool {
-    return left.baseURL == right.baseURL
-}
-
-/// HTTP environment.
-public protocol Environment {
-
-    /// Environment name (e.g. "QA", "Production")
-    var name: String { get }
-
-    /// Environment base URL.
-    var baseURL: URL { get }
-}
